requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
MarkupSafe==2.1.3
Flask-PageDown==0.4.0
WTForms==3.1.1
Flask-WTF==1.2.1
python-dotenv==0.19.0
werkzeug==3.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
tqdm==4.66.1
urllib3==2.1.0
celery==5.3.6
gunicorn==21.2.0
redis==3.5.3
email-validator==2.0.0